NEWS_NFTN_GADGETS_FEATURE/requirements.txt
Flask>=2.2.5
Flask-RESTful>=0.3.9
Jinja2>=3.1.3
Werkzeug>=2.3.7
requests==2.31.0
python-dotenv==1.0.0
